refactor: move comments and favorites api & queries to dedicated features
