+++ apps/frontend/src/features/recipes/recipes.queries.ts
@@ -1,10 +1,4 @@
-import type {
-  CreateCommentBody,
-  CreateRecipeBody,
-  PaginationQuery,
-  Recipe,
-  UpdateRecipeBody,
-} from "@recipes/shared";
+import type { CreateRecipeBody, UpdateRecipeBody } from "@recipes/shared";
 import {
   useInfiniteQuery,
   useMutation,
@@ -15,34 +9,21 @@ import type { MaybeRef } from "vue";
 import { toValue } from "vue";
 import type { RecipeFilters } from "./recipes.api";
 import {
-  addFavorite,
   createRecipe,
-  createRecipeComment,
-  deleteComment,
   deleteRecipe,
   getRecipe,
-  getRecipeComments,
   getRecipes,
-  removeFavorite,
   updateRecipe,
 } from "./recipes.api";
 
-const recipeKeys = {
+export const recipeKeys = {
   all: ["recipes"] as const,
   lists: () => [...recipeKeys.all, "list"] as const,
   list: (query: RecipeFilters) => [...recipeKeys.lists(), query] as const,
   detail: (id: string) => [...recipeKeys.all, id] as const,
   infinite: (query: RecipeFilters) =>
     [...recipeKeys.list(query), "infinite"] as const,
-
-  comments: {
-    all: ["comments"] as const,
-    lists: (id: string) =>
-      [...recipeKeys.detail(id), ...recipeKeys.comments.all] as const,
-    list: (id: string, query: PaginationQuery) =>
-      [...recipeKeys.comments.lists(id), query] as const,
-  },
-};
+} as const;
 
 export function useRecipes(filters: MaybeRef<RecipeFilters>) {
   return useQuery({
@@ -107,75 +88,6 @@ export function useDeleteRecipe() {
     onSuccess: (_, id) => {
       queryClient.removeQueries({ queryKey: recipeKeys.detail(id) });
       queryClient.invalidateQueries({ queryKey: recipeKeys.all });
-    },
-  });
-}
-
-export type ToggleFavoriteParams = {
-  id: string;
-  favorited: boolean;
-};
-
-export function useToggleFavorite() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: ({
-      id,
-      favorited,
-    }: ToggleFavoriteParams): Promise<{ favorited: boolean }> =>
-      favorited ? removeFavorite(id) : addFavorite(id),
-
-    onSuccess: (_, { id, favorited }) => {
-      queryClient.setQueryData<Recipe>(recipeKeys.detail(id), (old) =>
-        old ? { ...old, isFavorited: !favorited } : old,
-      );
-      queryClient.invalidateQueries({ queryKey: recipeKeys.all });
-    },
-  });
-}
-
-export function useRecipeComments(
-  id: MaybeRef<string>,
-  page: MaybeRef<number> = 1,
-  limit = 20,
-) {
-  const recipeId = toValue(id);
-  const query = { page: toValue(page), limit };
-
-  return useQuery({
-    queryKey: recipeKeys.comments.list(recipeId, query),
-    queryFn: () => getRecipeComments(recipeId, query),
-    enabled: () => !!recipeId,
-  });
-}
-
-export function useCreateComment() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: ({
-      recipeId,
-      body,
-    }: {
-      recipeId: string;
-      body: CreateCommentBody;
-    }) => createRecipeComment(recipeId, body),
-    onSuccess: (_, { recipeId }) => {
-      queryClient.invalidateQueries({
-        queryKey: recipeKeys.comments.lists(recipeId),
-      });
-    },
-  });
-}
-
-export function useDeleteComment() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: (commentId: string) => deleteComment(commentId),
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: recipeKeys.comments.all });
     },
   });
 }
